# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Twixtor Archive API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      parameters:
        - name: dramaId
          in: query
          schema: { type: integer }
        - name: actorId
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema:
            type: string
            enum: [slomo, non_slomo]
        - name: resolution
          in: query
          schema: { type: string }
        - name: fps
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: tags
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoListResponse"
    post:
      operationId: createVideo
      tags: [videos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoDetail"
    patch:
      operationId: updateVideo
      tags: [videos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
    delete:
      operationId: deleteVideo
      tags: [videos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /videos/{id}/view:
    post:
      operationId: recordView
      tags: [videos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
  /videos/{id}/download:
    post:
      operationId: downloadVideo
      tags: [videos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadResponse"
  /videos/{id}/rate:
    post:
      operationId: rateVideo
      tags: [videos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    Video:
      type: object
      required: [id, title, type, status]
      properties:
        id: { type: integer }
        title: { type: string }
        dramaId: { type: ["integer", "null"] }
        actorId: { type: ["integer", "null"] }
        dramaName: { type: ["string", "null"] }
        actorName: { type: ["string", "null"] }
        episode: { type: ["string", "null"] }
        scene: { type: ["string", "null"] }
        videoUrl: { type: ["string", "null"] }
        thumbnailUrl: { type: ["string", "null"] }
        type: { type: string }
        status: { type: string }
        resolution: { type: ["string", "null"] }
        fps: { type: ["integer", "null"] }
        duration: { type: ["integer", "null"] }
        fileSize: { type: ["integer", "null"] }
        format: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        viewCount: { type: integer }
        downloadCount: { type: integer }
        favoriteCount: { type: integer }
        popularityScore: { type: ["number", "null"] }
        isFavorited: { type: ["boolean", "null"] }
        isBookmarked: { type: ["boolean", "null"] }
        createdAt: { type: string }
    VideoDetail:
      type: object
      required: [id, title, type, status]
      properties:
        id: { type: integer }
        title: { type: string }
        dramaId: { type: ["integer", "null"] }
        actorId: { type: ["integer", "null"] }
        dramaName: { type: ["string", "null"] }
        actorName: { type: ["string", "null"] }
        episode: { type: ["string", "null"] }
        scene: { type: ["string", "null"] }
        videoUrl: { type: ["string", "null"] }
        thumbnailUrl: { type: ["string", "null"] }
        type: { type: string }
        status: { type: string }
        resolution: { type: ["string", "null"] }
        fps: { type: ["integer", "null"] }
        duration: { type: ["integer", "null"] }
        fileSize: { type: ["integer", "null"] }
        format: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        viewCount: { type: integer }
        downloadCount: { type: integer }
        favoriteCount: { type: integer }
        popularityScore: { type: ["number", "null"] }
        isFavorited: { type: ["boolean", "null"] }
        isBookmarked: { type: ["boolean", "null"] }
        createdAt: { type: string }
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    VideoListResponse:
      type: object
      required: [videos, total, page, limit]
      properties:
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    VideoInput:
      type: object
      required: [title, type]
      properties:
        title: { type: string }
        dramaId: { type: integer }
        actorId: { type: integer }
        episode: { type: string }
        scene: { type: string }
        videoUrl: { type: string }
        thumbnailUrl: { type: string }
        type: { type: string, enum: [slomo, non_slomo] }
        status: { type: string }
        resolution: { type: string }
        fps: { type: integer }
        tags:
          type: array
          items: { type: string }
    VideoUpdate:
      type: object
      properties:
        title: { type: string }
        dramaId: { type: integer }
        actorId: { type: integer }
        episode: { type: string }
        scene: { type: string }
        videoUrl: { type: string }
        thumbnailUrl: { type: string }
        type: { type: string }
        status: { type: string }
        resolution: { type: string }
        fps: { type: integer }
        tags:
          type: array
          items: { type: string }
...
    DownloadResponse:
      type: object
      required: [url]
      properties:
        url: { type: string }
    RatingInput:
      type: object
      required: [rating]
      properties:
        rating: { type: integer }
    Comment:
      type: object
      required: [id, text]
      properties:
        id: { type: integer }
        videoId: { type: integer }
        userId: { type: integer }
        username: { type: string }
        text: { type: string }
        parentId: { type: ["integer", "null"] }
        likeCount: { type: integer }
        replies: { type: array, items: { $ref: "#/components/schemas/Comment" } }
        createdAt: { type: string }
...